🚧 Restore non web compatibility
