+++ lib/infrastructure/rpc/browser_extension_aws.nonweb.dart
@@ -1,0 +1,13 @@
+import 'dart:async';
+
+class BrowserExtensionAWS {
+  BrowserExtensionAWS() {
+    throw UnimplementedError(
+      'Browser extension is not supported on the desktop platform.',
+    );
+  }
+
+  static bool get isPlatformCompatible => false;
+
+  Future<void> run() async {}
+}
